refactor: chessBoard 패키지 생성
