+++ src/main/java/chess/chessBoard/ChessBoard.java
@@ -1,4 +1,4 @@
-package chess;
+package chess.chessBoard;
 
 import java.util.List;
