+++ src/main/java/chess/chessBoard/Lettering.java
@@ -1,4 +1,4 @@
-package chess;
+package chess.chessBoard;
 
 public enum Lettering {
