+++ src/main/java/chess/chessBoard/Numbering.java
@@ -1,4 +1,4 @@
-package chess;
+package chess.chessBoard;
 
 public enum Numbering {
